re-organizing codecs into companion objects (to help with compile time)
