+++ src/main/scala/hydrozoa/config/head/parameters/HeadParameters.scala
@@ -4,9 +4,13 @@ import hydrozoa.config.head.multisig.fallback.FallbackContingency
 import hydrozoa.config.head.multisig.settlement.SettlementConfig
 import hydrozoa.config.head.multisig.timing.TxTiming
 import hydrozoa.config.head.multisig.timing.TxTiming.Durations.{DepositAbsorptionDuration, DepositMaturityDuration, DepositSubmissionDuration, InactivityMarginDuration, MinSettlementDuration, SilenceDuration}
+import hydrozoa.config.head.network.CardanoNetwork
 import hydrozoa.config.head.rulebased.dispute.DisputeResolutionConfig
+import hydrozoa.lib.cardano.cip116.JsonCodecs.CIP0116.Conway.given
 import hydrozoa.lib.cardano.scalus.QuantizedTime.QuantizedFiniteDuration
 import hydrozoa.lib.number.PositiveInt
+import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
+import io.circe.{Decoder, Encoder}
 import scalus.cardano.ledger.Hash32
 
 /** The parameters that peers agree upon to run the protocol. These parameters get hashed into the
@@ -29,6 +33,12 @@ final case class HeadParameters(
 }
 
 object HeadParameters {
+
+    given headParametersEncoder: Encoder[HeadParameters] = deriveEncoder[HeadParameters]
+
+    given headParametersDecoder(using CardanoNetwork.Section): Decoder[HeadParameters] =
+        deriveDecoder[HeadParameters]
+
     trait Section
         extends TxTiming.Section,
           FallbackContingency.Section,
