+++ src/main/scala/hydrozoa/config/head/rulebased/dispute/DisputeResolutionConfig.scala
@@ -1,6 +1,9 @@
 package hydrozoa.config.head.rulebased.dispute
 
+import hydrozoa.config.head.network.CardanoNetwork
 import hydrozoa.lib.cardano.scalus.QuantizedTime.QuantizedFiniteDuration
+import io.circe.*
+import io.circe.generic.semiauto.*
 import scala.concurrent.duration.DurationInt
 import scalus.cardano.ledger.SlotConfig
 
@@ -22,4 +25,12 @@ object DisputeResolutionConfig {
         DisputeResolutionConfig(
           QuantizedFiniteDuration(slotConfig = slotConfig, finiteDuration = 2.days)
         )
+
+    given disputeResolutionConfigEncoder: Encoder[DisputeResolutionConfig] =
+        deriveEncoder[DisputeResolutionConfig]
+
+    given disputeResolutionConfigDecoder(using
+        CardanoNetwork.Section
+    ): Decoder[DisputeResolutionConfig] =
+        deriveDecoder[DisputeResolutionConfig]
 }
