+++ src/main/scala/hydrozoa/config/node/operation/multisig/NodeOperationMultisigConfig.scala
@@ -1,6 +1,9 @@
 package hydrozoa.config.node.operation.multisig
 
+import hydrozoa.config.head.multisig.timing.given
 import hydrozoa.lib.number.PositiveInt
+import io.circe.*
+import io.circe.generic.semiauto.*
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
 
 final case class NodeOperationMultisigConfig(
@@ -22,4 +25,8 @@ object NodeOperationMultisigConfig {
       cardanoLiaisonPollingPeriod = 10.seconds,
       peerLiaisonMaxEventsPerBatch = PositiveInt.unsafeApply(42)
     )
+
+    given Encoder[NodeOperationMultisigConfig] = deriveEncoder[NodeOperationMultisigConfig]
+
+    given Decoder[NodeOperationMultisigConfig] = deriveDecoder[NodeOperationMultisigConfig]
 }
