+++ src/main/scala/hydrozoa/config/node/owninfo/OwnHeadPeerPrivate.scala
@@ -1,7 +1,11 @@
 package hydrozoa.config.node.owninfo
 
 import hydrozoa.config.head.peers.HeadPeers
+import hydrozoa.multisig.consensus.peer.HeadPeerWallet.dummyHeadPeerWalletEncoder
 import hydrozoa.multisig.consensus.peer.{HeadPeerId, HeadPeerNumber, HeadPeerWallet}
+import io.circe.*
+import io.circe.generic.semiauto.*
+import io.circe.syntax.*
 import scalus.crypto.ed25519.VerificationKey
 
 final case class OwnHeadPeerPrivate private (
@@ -29,4 +33,30 @@ object OwnHeadPeerPrivate {
         override def ownHeadPeerNum: HeadPeerNumber = ownHeadPeerPublic.ownHeadPeerNum
         override def ownHeadVKey: VerificationKey = ownHeadPeerPublic.ownHeadVKey
     }
+
+    given dummyOwnHeadPeerPrivateEncoder: Encoder[OwnHeadPeerPrivate] =
+        Encoder.instance(ownHeadPeerPrivate =>
+            Json.obj(
+              "ownHeadWallet" -> ownHeadPeerPrivate.ownHeadWallet.asJson(using
+                dummyHeadPeerWalletEncoder
+              ),
+            )
+        )
+
+    given ownHeadPeerPrivateDecoder(using
+        peers: HeadPeers.Section
+    ): Decoder[OwnHeadPeerPrivate] = Decoder
+        .instance(c =>
+            for {
+                wallet <- c.downField("ownHeadWallet").as[HeadPeerWallet]
+            } yield OwnHeadPeerPrivate(wallet, peers.headPeers)
+        )
+        .emap {
+            case Some(ownHeadPeerPrivate: OwnHeadPeerPrivate) => Right(ownHeadPeerPrivate)
+            case None =>
+                Left(
+                  "Could not construct head peer private section. Does the wallet's peer number correspond to " +
+                      "an existing HeadPeerNumber?"
+                )
+        }
 }
