+++ src/main/scala/hydrozoa/lib/number/NonNegativeInt.scala
@@ -1,5 +1,7 @@
 package hydrozoa.lib.number
 
+import io.circe.*
+
 type NonNegativeInt = NonNegativeInt.NonNegativeInt
 
 object NonNegativeInt {
@@ -13,4 +15,13 @@ object NonNegativeInt {
     }
 
     given Conversion[NonNegativeInt, scala.Int] = identity
+
+    given nonNegativeIntEncoder: Encoder[NonNegativeInt] = Encoder.encodeInt
+
+    given nonNegativeIntDecoder: Decoder[NonNegativeInt] = Decoder.decodeInt.emap { i =>
+        NonNegativeInt.apply(i) match {
+            case None      => Left(s"Expected a non-negative integer, got $i")
+            case Some(nni) => Right(nni)
+        }
+    }
 }
