+++ src/main/scala/hydrozoa/multisig/consensus/peer/HeadPeerWallet.scala
@@ -1,10 +1,15 @@
 package hydrozoa.multisig.consensus.peer
 
+import hydrozoa.lib.cardano.cip116.JsonCodecs.CIP0116.Conway.given
+import hydrozoa.lib.cardano.scalus.codecs.json.Codecs.dummySigningKey
 import hydrozoa.lib.cardano.scalus.txbuilder.Transaction.attachVKeyWitnesses
 import hydrozoa.lib.cardano.wallet.*
 import hydrozoa.multisig.consensus.ack.{AckBlock, AckId, AckNumber}
+import hydrozoa.multisig.consensus.peer.HeadPeerNumber.given
 import hydrozoa.multisig.ledger.block.{Block, BlockHeader}
 import hydrozoa.multisig.ledger.l1.tx.TxSignature
+import io.circe.*
+import io.circe.syntax.*
 import scala.language.implicitConversions
 import scalus.cardano.ledger.{Transaction, VKeyWitness}
 import scalus.crypto.ed25519.{SigningKey, VerificationKey}
@@ -130,4 +135,29 @@ object HeadPeerWallet:
       walletModule = WalletModule.Scalus,
       verificationKey = verificationKey,
       signingKey = signingKey
+    )
+
+    /** We can't directly access the signing key of a head peer wallet, so we use a "dummy"
+      * all-zeros signing key instead
+      * @return
+      */
+    given dummyHeadPeerWalletEncoder: Encoder[HeadPeerWallet] = new Encoder[HeadPeerWallet] {
+        override def apply(w: HeadPeerWallet): Json = Json.obj(
+          "peerNum" -> w.getPeerNum.asJson,
+          "verificationKey" -> w.exportVerificationKey.asJson(using given_Encoder_VerificationKey),
+          "signingKey" -> dummySigningKey.asJson
+        )
+    }
+
+    given headPeerWalletDecoder: Decoder[HeadPeerWallet] = Decoder.instance(c =>
+        for {
+            peerNum <- c.downField("peerNum").as[HeadPeerNumber]
+            vkey <- c.downField("verificationKey").as[VerificationKey]
+            skey <- c.downField("signingKey").as[SigningKey]
+            w = HeadPeerWallet.scalusWallet(
+              peerNum = peerNum,
+              verificationKey = vkey,
+              signingKey = skey
+            )
+        } yield w
     )
